backend/requirements.txt
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.104.1
fonttools==4.59.0
fuzzywuzzy==0.18.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.23.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
networkx==3.5
numpy==1.26.4
opencv-python==4.8.1.78
packaging==25.0
pandas==2.1.4
passlib==1.7.4
pdf2image==1.16.3
Pillow==10.1.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.2.3
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-Levenshtein==0.23.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2023.10.3
reportlab==4.4.3
requests==2.32.4
rsa==4.9.1
scikit-image==0.22.0
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==9.0.0
tifffile==2025.6.11
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn[standard]==0.24.0
watchfiles==1.1.0
websockets==15.0.1
pytest==8.4.1



# NER un AI bibliotēkas (opsionālas - tiks pievienotas nākotnē)
# spacy>=3.7.0  # NER un NLP
# spacy-transformers>=1.3.0  # BERT transformers
# torch>=2.0.0  # PyTorch for deep learning models
# transformers>=4.35.0  # Hugging Face transformers
# datasets>=2.15.0  # Dataset handling
